requirements.txt
highrise-bot-sdk>=24.1.0
Flask==3.0.2
requests==2.31.0
gunicorn==21.2.0
pydub
yt-dlp
tenacity==5.1.5
mega.py==1.0.8